+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -187,6 +187,17 @@ export const configSchema = Joi.object({
     headers: Joi.object().pattern(Joi.string(), Joi.string()),
     overrideBaseURL: Joi.string().uri().optional(),
   }).optional(),
+  cdnLogsConfig: Joi.object({
+    bucketName: Joi.string().required(),
+    filters: Joi.array().items(
+      Joi.object({
+        key: Joi.string().required(),
+        value: Joi.array().items(Joi.string()).required(),
+        type: Joi.string().valid('include', 'exclude').optional(),
+      }),
+    ).optional(),
+    outputLocation: Joi.string().required(),
+  }).optional(),
   contentAiConfig: Joi.object({
     index: Joi.string().optional(),
   }).optional(),
@@ -252,6 +263,7 @@ export const Config = (data = {}) => {
   self.getLatestMetrics = (type) => state?.handlers?.[type]?.latestMetrics;
   self.getFetchConfig = () => state?.fetchConfig;
   self.getBrandConfig = () => state?.brandConfig;
+  self.getCdnLogsConfig = () => state?.cdnLogsConfig;
 
   self.updateSlackConfig = (channel, workspace, invitedUserCount) => {
     state.slack = {
@@ -353,6 +365,10 @@ export const Config = (data = {}) => {
     return config?.enabled ?? false;
   };
 
+  self.updateCdnLogsConfig = (cdnLogsConfig) => {
+    state.cdnLogsConfig = cdnLogsConfig;
+  };
+
   return Object.freeze(self);
 };
 
@@ -365,4 +381,5 @@ Config.toDynamoItem = (config) => ({
   imports: config.getImports(),
   fetchConfig: config.getFetchConfig(),
   brandConfig: config.getBrandConfig(),
+  cdnLogsConfig: config.getCdnLogsConfig(),
 });