+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@adobe/spacecat-shared-google-client-v1.4.36](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.35...@adobe/spacecat-shared-google-client-v1.4.36) (2025-07-05)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#837](https://github.com/adobe/spacecat-shared/issues/837)) ([acfbbe7](https://github.com/adobe/spacecat-shared/commit/acfbbe712d90fe2f4b0cab97e8a941eb1bd5c8ea))
+
+# [@adobe/spacecat-shared-google-client-v1.4.35](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.34...@adobe/spacecat-shared-google-client-v1.4.35) (2025-06-28)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#830](https://github.com/adobe/spacecat-shared/issues/830)) ([b98589d](https://github.com/adobe/spacecat-shared/commit/b98589da5c11aa4e63358e98f0c0852b0ef2a02d))
+
+# [@adobe/spacecat-shared-google-client-v1.4.34](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.33...@adobe/spacecat-shared-google-client-v1.4.34) (2025-06-21)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#817](https://github.com/adobe/spacecat-shared/issues/817)) ([522b440](https://github.com/adobe/spacecat-shared/commit/522b440ac1ac1d704b4106ad1f90f192541a8f0c))
+
 # [@adobe/spacecat-shared-google-client-v1.4.33](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.32...@adobe/spacecat-shared-google-client-v1.4.33) (2025-06-14)